+++ src/main/java/org/prebid/server/hooks/execution/v1/exitpoint/ExitpointPayloadImpl.java
@@ -3,7 +3,7 @@ package org.prebid.server.hooks.execution.v1.exitpoint;
 import io.vertx.core.MultiMap;
 import lombok.Value;
 import lombok.experimental.Accessors;
-import org.prebid.server.hooks.v1.exit.ExitpointPayload;
+import org.prebid.server.hooks.v1.exitpoint.ExitpointPayload;
 
 @Accessors(fluent = true)
 @Value(staticConstructor = "of")
